third_party/WebKit/LayoutTests/fast/dom/setter-type-enforcement.html

<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>

<script>
description("Tests type enforcement on DOM setters.");

// NodeList is not a Node, so document.body can't be set to one.
var nodelist = document.getElementsByName('sillypants');
shouldThrow("document.body = nodelist;");
</script>
</body>
</html>
